+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/source/DefaultScopeDefine.java
@@ -100,6 +100,7 @@ public class DefaultScopeDefine {
     public static final int EBPF_PROFILING_TASK = 46;
     public static final int EBPF_PROFILING_SCHEDULE = 47;
     public static final int EBPF_PROFILING_DATA = 48;
+    public static final int PROCESS_SERVICE_LABEL = 49;
 
     /**
      * Catalog of scope, the metrics processor could use this to group all generated metrics by oal rt.