+++ oap-server/server-receiver-plugin/skywalking-ebpf-receiver-plugin/src/main/java/org/apache/skywalking/oap/server/receiver/ebpf/provider/handler/EBPFProfilingServiceHandler.java
@@ -36,12 +36,12 @@ import org.apache.skywalking.oap.server.core.source.EBPFProfilingData;
 import org.apache.skywalking.oap.server.core.source.EBPFProcessProfilingSchedule;
 import org.apache.skywalking.oap.server.core.source.SourceReceiver;
 import org.apache.skywalking.oap.server.core.storage.StorageModule;
-import org.apache.skywalking.oap.server.core.storage.profiling.ebpf.EBPFProfilingProcessFinder;
 import org.apache.skywalking.oap.server.core.storage.profiling.ebpf.IEBPFProfilingTaskDAO;
 import org.apache.skywalking.oap.server.core.storage.query.IMetadataQueryDAO;
 import org.apache.skywalking.oap.server.library.module.ModuleManager;
 import org.apache.skywalking.oap.server.library.server.grpc.GRPCHandler;
 import org.apache.skywalking.oap.server.library.util.CollectionUtils;
+import org.apache.skywalking.oap.server.network.trace.component.command.EBPFProfilingTaskCommand;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
@@ -49,6 +49,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 /**
@@ -85,11 +86,12 @@ public class EBPFProfilingServiceHandler extends EBPFProfilingServiceGrpc.EBPFPr
             }
 
             // fetch tasks from process id list
-            final EBPFProfilingProcessFinder finder = EBPFProfilingProcessFinder.builder()
-                    .processIdList(processes.stream().map(Process::getId).collect(Collectors.toList())).build();
-            final List<EBPFProfilingTask> tasks = taskDAO.queryTasks(finder, null, 0, latestUpdateTime);
+            final List<String> serviceIdList = processes.stream().map(Process::getServiceId).distinct().collect(Collectors.toList());
+            final List<EBPFProfilingTask> tasks = taskDAO.queryTasks(serviceIdList, null, 0, latestUpdateTime);
+
             final Commands.Builder builder = Commands.newBuilder();
-            tasks.stream().map(t -> commandService.newEBPFProfilingTaskCommand(t).serialize()).forEach(builder::addCommands);
+            tasks.stream().flatMap(t -> this.buildProfilingCommands(t, processes).stream())
+                    .map(EBPFProfilingTaskCommand::serialize).forEach(builder::addCommands);
             responseObserver.onNext(builder.build());
             responseObserver.onCompleted();
             return;
@@ -98,6 +100,35 @@ public class EBPFProfilingServiceHandler extends EBPFProfilingServiceGrpc.EBPFPr
         }
         responseObserver.onNext(Commands.newBuilder().build());
         responseObserver.onCompleted();
+    }
+
+    private List<EBPFProfilingTaskCommand> buildProfilingCommands(EBPFProfilingTask task, List<Process> processes) {
+        final ArrayList<EBPFProfilingTaskCommand> commands = new ArrayList<>(processes.size());
+        for (Process process : processes) {
+            // The service id must matches between process and task
+            if (!Objects.equals(process.getServiceId(), task.getServiceId())) {
+                continue;
+            }
+
+            // If the task haven't labeled requirement, them just add the process
+            boolean shouldAdd = false;
+            if (CollectionUtils.isEmpty(task.getProcessLabels())) {
+                shouldAdd = true;
+            } else {
+                // If the process have matched one of the task label, them need to add the command
+                for (String label : task.getProcessLabels()) {
+                    if (process.getLabels().contains(label)) {
+                        shouldAdd = true;
+                        break;
+                    }
+                }
+            }
+
+            if (shouldAdd) {
+                commands.add(commandService.newEBPFProfilingTaskCommand(task, process.getId()));
+            }
+        }
+        return commands;
     }
 
     @Override