+++ oap-server/server-storage-plugin/storage-jdbc-hikaricp-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/jdbc/h2/dao/H2EBPFProfilingTaskDAO.java
@@ -18,17 +18,15 @@
 
 package org.apache.skywalking.oap.server.storage.plugin.jdbc.h2.dao;
 
+import com.google.gson.Gson;
 import lombok.AllArgsConstructor;
 import org.apache.skywalking.oap.server.core.analysis.IDManager;
-import org.apache.skywalking.oap.server.core.query.type.EBPFProfilingProcessFinderType;
 import org.apache.skywalking.oap.server.core.profiling.ebpf.storage.EBPFProfilingTargetType;
 import org.apache.skywalking.oap.server.core.profiling.ebpf.storage.EBPFProfilingTaskRecord;
 import org.apache.skywalking.oap.server.core.profiling.ebpf.storage.EBPFProfilingTriggerType;
 import org.apache.skywalking.oap.server.core.query.type.EBPFProfilingTask;
-import org.apache.skywalking.oap.server.core.storage.profiling.ebpf.EBPFProfilingProcessFinder;
 import org.apache.skywalking.oap.server.core.storage.profiling.ebpf.IEBPFProfilingTaskDAO;
 import org.apache.skywalking.oap.server.library.client.jdbc.hikaricp.JDBCHikariCPClient;
-import org.apache.skywalking.oap.server.library.util.CollectionUtils;
 import org.apache.skywalking.oap.server.library.util.StringUtil;
 
 import java.io.IOException;
@@ -36,36 +34,23 @@ import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 @AllArgsConstructor
 public class H2EBPFProfilingTaskDAO implements IEBPFProfilingTaskDAO {
+    private static final Gson GSON = new Gson();
     private JDBCHikariCPClient h2Client;
 
     @Override
-    public List<EBPFProfilingTask> queryTasks(EBPFProfilingProcessFinder finder, EBPFProfilingTargetType targetType, long taskStartTime, long latestUpdateTime) throws IOException {
+    public List<EBPFProfilingTask> queryTasks(List<String> serviceIdList, EBPFProfilingTargetType targetType, long taskStartTime, long latestUpdateTime) throws IOException {
         final StringBuilder sql = new StringBuilder();
         List<Object> condition = new ArrayList<>(7);
         sql.append("select * from ").append(EBPFProfilingTaskRecord.INDEX_NAME);
 
         StringBuilder conditionSql = new StringBuilder();
 
-        if (finder.getFinderType() != null) {
-            appendCondition(conditionSql, condition,
-                    EBPFProfilingTaskRecord.PROCESS_FIND_TYPE, finder.getFinderType().value());
-        }
-        if (StringUtil.isNotEmpty(finder.getServiceId())) {
-            appendCondition(conditionSql, condition,
-                    EBPFProfilingTaskRecord.SERVICE_ID, finder.getServiceId());
-        }
-        if (StringUtil.isNotEmpty(finder.getInstanceId())) {
-            appendCondition(conditionSql, condition,
-                    EBPFProfilingTaskRecord.INSTANCE_ID, finder.getInstanceId());
-        }
-        if (CollectionUtils.isNotEmpty(finder.getProcessIdList())) {
-            appendListCondition(conditionSql, condition,
-                    EBPFProfilingTaskRecord.PROCESS_ID, finder.getProcessIdList());
-        }
+        appendListCondition(conditionSql, condition, EBPFProfilingTaskRecord.SERVICE_ID, serviceIdList);
         if (targetType != null) {
             appendCondition(conditionSql, condition,
                     EBPFProfilingTaskRecord.TARGET_TYPE, targetType.value());
@@ -98,16 +83,15 @@ public class H2EBPFProfilingTaskDAO implements IEBPFProfilingTaskDAO {
         while (resultSet.next()) {
             EBPFProfilingTask task = new EBPFProfilingTask();
             task.setTaskId(resultSet.getString(H2TableInstaller.ID_COLUMN));
-            task.setProcessFinderType(EBPFProfilingProcessFinderType.valueOf(
-                    resultSet.getInt(EBPFProfilingTaskRecord.PROCESS_FIND_TYPE)));
             final String serviceId = resultSet.getString(EBPFProfilingTaskRecord.SERVICE_ID);
             task.setServiceId(serviceId);
             task.setServiceName(IDManager.ServiceID.analysisId(serviceId).getName());
-            final String instanceId = resultSet.getString(EBPFProfilingTaskRecord.INSTANCE_ID);
-            task.setInstanceId(instanceId);
-            task.setInstanceName(IDManager.ServiceInstanceID.analysisId(instanceId).getName());
-            task.setProcessId(resultSet.getString(EBPFProfilingTaskRecord.PROCESS_ID));
-            task.setProcessName(resultSet.getString(EBPFProfilingTaskRecord.PROCESS_NAME));
+            final String processLabelString = resultSet.getString(EBPFProfilingTaskRecord.PROCESS_LABELS_JSON);
+            if (StringUtil.isNotEmpty(processLabelString)) {
+                task.setProcessLabels(GSON.<List<String>>fromJson(processLabelString, ArrayList.class));
+            } else {
+                task.setProcessLabels(Collections.emptyList());
+            }
             task.setTaskStartTime(resultSet.getLong(EBPFProfilingTaskRecord.START_TIME));
             task.setTriggerType(EBPFProfilingTriggerType.valueOf(
                     resultSet.getInt(EBPFProfilingTaskRecord.TRIGGER_TYPE)));